+++ test/0.8.9/contracts/TriggerableWithdrawalsGateway__MockForVEB.sol
@@ -1,7 +1,7 @@
 pragma solidity 0.8.9;
 
 contract TriggerableWithdrawalsGateway__MockForVEB {
-    event Mock__triggerFullWithdrawalsTriggered(uint256 exitsCount, address refundRecipient, uint8 exitType);
+    event Mock__triggerFullWithdrawalsTriggered(uint256 exitsCount, address refundRecipient, uint256 exitType);
 
     struct ValidatorData {
         uint256 stakingModuleId;
@@ -12,7 +12,7 @@ contract TriggerableWithdrawalsGateway__MockForVEB {
     function triggerFullWithdrawals(
         ValidatorData[] calldata triggerableExitData,
         address refundRecipient,
-        uint8 exitType
+        uint256 exitType
     ) external payable {
         emit Mock__triggerFullWithdrawalsTriggered(triggerableExitData.length, refundRecipient, exitType);
     }
